+++ redwood-composeui/src/commonMain/kotlin/app/cash/redwood/composeui/RedwoodContent.kt
@@ -90,7 +90,7 @@ public fun RedwoodContent(
       }
     },
   ) {
-    children.render()
+    children.Render()
   }
 }